Package Tasks and Add Javadocs
