--- docs/sampleInteraction.txt
@@ -1,50 +1,0 @@
-____________________________________________________________
-<Logo>
-Hello! I'm Duke
-What can I do for you?
-____________________________________________________________
-
-list
-
-____________________________________________________________
-Here are the tasks in your list:
-1.[T][X] read book
-2.[D][ ] return book (by: June 6th)
-3.[E][ ] project meeting (at: Aug 6th 2-4pm)
-4.[T][X] join sports club
-____________________________________________________________
-
-todo borrow book
-
-____________________________________________________________
-Got it. I've added this task:
-[T][ ] borrow book
-Now you have 5 tasks in the list.
-____________________________________________________________
-
-deadline return book /by Sunday
-
-____________________________________________________________
-Got it. I've added this task:
-[D][ ] return book (by: Sunday)
-Now you have 6 tasks in the list.
-____________________________________________________________
-
-mark 2
-
-____________________________________________________________
-Nice! I've marked this task as done:
-[D][X] return book (by: June 6th)
-____________________________________________________________
-
-blah
-
-____________________________________________________________
-☹ OOPS!!! I'm sorry, but I don't know what that means :-(
-____________________________________________________________
-
-bye
-
-____________________________________________________________
-Bye. Hope to see you again soon!
-____________________________________________________________
