--- src/main/java/Deadline.java
@@ -1,16 +1,0 @@
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-
-public class Deadline extends Task {
-    protected LocalDate by;
-
-    public Deadline(String s, String by) throws LilyException {
-        super(s);
-        this.by = LocalDate.parse(by);
-    }
-
-    @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (by: " + by.format(DateTimeFormatter.ofPattern("dd MMM yy")) + ")";
-    }
-}
