+++ src/main/java/lily/task/LilyException.java
@@ -1,0 +1,18 @@
+package lily.task;
+
+/**
+ * Exceptions thrown by Lily-related functions
+ * 
+ * Author: Hong Yi En, Ian
+ * Last Updated: Jan 2022 (AY21/22 Sem 2)
+ */
+public class LilyException extends Exception {
+    public LilyException(String msg) {
+        super(msg);
+    }
+
+    @Override
+    public String toString() {
+        return "error message bro";
+    }
+}
