+++ src/main/java/lily/task/Task.java
@@ -1,34 +1,47 @@
+package lily.task;
 import java.io.Serializable;
 
+/**
+ * A parent class which Todo, Events and Deadlines derive from
+ * 
+ * @author Hong Yi En, Ian
+ * @version Jan 2022 (AY21/22 Sem 2)
+ */
 public class Task implements Serializable{
     protected String desc;    
     protected boolean isDone;    
 
+    /**
+     * Create a Task.
+     * @param desc What the Task is about.
+     */
     public Task(String desc) {
         this.desc = desc;
         this.isDone = false;
     }
 
+    /**
+     * Mark the task as done.
+     */
     public void mark() {
         this.isDone = true;
     }
 
+    /**
+     * Mark the task as undone.
+     */
     public void unmark() {
         this.isDone = false;
     }
 
-    public String getDesc() {
-        return this.desc;
-    }
-
-    public boolean isThisDone() {
-        return this.isDone;
-    }
-
-    public String getStatusIcon() {
+    private String getStatusIcon() {
         return "[" + (isDone ? "X" : " ") + "] ";
     }
 
+    /**
+     * Returns the Task's description and whether it is done
+     * @return Task in the form of "[ ] Description"
+     */
     @Override
     public String toString() {
         return this.getStatusIcon() + this.desc;
